Changed directory of moved pixel code to be CondFormats/SiPixelTransient/
