+++ CondFormats/SiPixelTransient/src/SiPixelGenError.cc
@@ -31,7 +31,7 @@
 
 
 #ifndef SI_PIXEL_TEMPLATE_STANDALONE
-#include "CondFormats/SiPixelCommon/interface/SiPixelGenError.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelGenError.h"
 #include "FWCore/ParameterSet/interface/FileInPath.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #define LOGERROR(x) LogError(x)
