+++ CondTools/SiPixel/BuildFile.xml
@@ -3,6 +3,7 @@
 <use   name="FWCore/MessageLogger"/>
 <use   name="CondFormats/DataRecord"/>
 <use   name="CondFormats/SiPixelObjects"/>
+<use   name="CondFormats/SiPixelTransient"/>
 <use   name="CondCore/PopCon"/>
 <export>
   <lib   name="1"/>
