+++ CondTools/SiPixel/test/SiPixelGenErrorDBObjectReader.cc
@@ -9,7 +9,7 @@
 
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/EventSetup.h"
-#include "CondFormats/SiPixelCommon/interface/SiPixelGenError.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelGenError.h"
 //#include "MagneticField/Engine/interface/MagneticField.h"
 //#include "FWCore/ParameterSet/interface/FileInPath.h"
