+++ FastSimulation/TrackingRecHitProducer/BuildFile.xml
@@ -28,7 +28,7 @@
 <use   name="SimDataFormats/CrossingFrame"/>
 <use   name="SimDataFormats/EncodedEventId"/>
 <use   name="SimDataFormats/TrackingHit"/>
-<use   name="CondFormats/SiPixelCommon"/>
+<use   name="CondFormats/SiPixelTransient"/>
 <use   name="boost"/>
 <use   name="clhep"/>
 <use   name="gsl"/>
