+++ RecoLocalTracker/SiPixelRecHits/interface/SiPixelTemplateReco.h
@@ -54,8 +54,8 @@
 #define SiPixelTemplateReco_h 1
 
 #ifndef SI_PIXEL_TEMPLATE_STANDALONE
-#include "CondFormats/SiPixelCommon/interface/SiPixelTemplateDefs.h"
-#include "CondFormats/SiPixelCommon/interface/SiPixelTemplate.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelTemplateDefs.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelTemplate.h"
 #else
 #include "SiPixelTemplateDefs.h"
 #include "SiPixelTemplate.h"
