+++ RecoLocalTracker/SiPixelRecHits/interface/SiPixelTemplateSplit.h
@@ -29,11 +29,11 @@
 #ifndef SiPixelTemplateSplit_h
 #define SiPixelTemplateSplit_h 1
 
-#include "CondFormats/SiPixelCommon/interface/SiPixelTemplateDefs.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelTemplateDefs.h"
 
 #ifndef SI_PIXEL_TEMPLATE_STANDALONE
-#include "CondFormats/SiPixelCommon/interface/SiPixelTemplate.h"
-#include "CondFormats/SiPixelCommon/interface/SiPixelTemplate2D.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelTemplate.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelTemplate2D.h"
 #else
 #include "SiPixelTemplate.h"
 #include "SiPixelTemplate2D.h"
