+++ SimTracker/SiPixelDigitizer/plugins/SiPixelDigitizerAlgorithm.h
@@ -14,7 +14,7 @@
 #include "DataFormats/Math/interface/approx_exp.h"
 #include "SimDataFormats/PileupSummaryInfo/interface/PileupMixingContent.h"
 #include "Geometry/CommonTopologies/interface/PixelTopology.h"
-#include "CondFormats/SiPixelCommon/interface/SiPixelTemplate2D.h"
+#include "CondFormats/SiPixelTransient/interface/SiPixelTemplate2D.h"
 #include "CondFormats/SiPixelObjects/interface/SiPixel2DTemplateDBObject.h"
 #include "boost/multi_array.hpp"
